requirements.txt
Django==2.1.5
django-enumchoicefield==1.0.2
django-filter==2.1.0
djangorestframework==3.9.1
Markdown==3.0.1
pytz==2018.9
